STABLE KERNEL VERSION 2.4.20 RELEASED
The kernel builds and works fine for me. not much patches for that  version yet available - only xfs at the time of this submit, but that might be allready enough for quite a lot of the users out there.
I recommend giving it a short period in Devel before pulling it to test, but i leave it up to erics deciding when to do so. (maybe when there also is preempt and some more available for it)


The Line in DETAILS is only for the scripts that parse the version out of there - so we get 2.4.20 shown on distrowatch