Version bump to 0.1.28.
I'm not moving it into test as of yet, because it hangs during the
creation of the MD5 file at the end of the cast.  I think it's an old
sorcery bug that has not yet been fixed, but if anyone cares to prove
me wrong, I'm OK with that.  :^)