Changed the behaviour of the firefox executable: now running `firefox` will connect to the current running firefox-bin and try to open up another window instead of giving you the profile manager. However you can still run multiple profiles by explicitly running `firefox -P <profile>`.
I don't know if this is desired behaviour or not, I find it much more
convienient to have a button attached to this now because I'm always
assured to get a new window even if I forget to close firefox on another
desktop. I sent an email to the sm-discuss about it.