Update to gnome 2.12.0 beta 2 if not a new version then just did the gpg conversion
Notes: 
iso-codes added as it is needed by epiphany and totem. I could only find
a tarball of it in the debian stuff.

dbus and hal needed to be updated for gnome 2.12.0 beta 2 to build

updated cairo to 0.9.0 as it is needed to build as well but cannot
update to 0.9.2 or gtk+ 2.8.0 until pango updated as cairo 0.9.2 removes
the libpixman depends

added libxft 2.1.7 as control-center2 needs it for building.  This probably
needs some major correcting to work with all the X11-LIBS as I just used
the installs --prefix=/usr/X11R6 and it replaces files installed by the 
X11-LIBS.